Moves `deploy_jar` target and supporting things to `pants.jvm`
# Rust tests and lints will be skipped. Delete if not intended.
[ci skip-rust]
