+++ src/python/pants/backend/experimental/java/register.py
@@ -5,9 +5,7 @@ from pants.backend.java.compile import javac
 from pants.backend.java.dependency_inference import java_parser, java_parser_launcher
 from pants.backend.java.dependency_inference import rules as dependency_inference_rules
 from pants.backend.java.goals import check, tailor
-from pants.backend.java.package import deploy_jar
 from pants.backend.java.target_types import (
-    DeployJarTarget,
     JavaSourcesGeneratorTarget,
     JavaSourceTarget,
     JunitTestsGeneratorTarget,
@@ -18,8 +16,9 @@ from pants.jvm import classpath, jdk_rules, resources
 from pants.jvm import util_rules as jvm_util_rules
 from pants.jvm.dependency_inference import symbol_mapper
 from pants.jvm.goals import lockfile
+from pants.jvm.package import deploy_jar
 from pants.jvm.resolve import coursier_fetch, jvm_tool
-from pants.jvm.target_types import JvmArtifactTarget
+from pants.jvm.target_types import DeployJarTarget, JvmArtifactTarget
 from pants.jvm.test import junit
