+++ src/python/pants/backend/experimental/scala/register.py
@@ -1,7 +1,5 @@
 # Copyright 2021 Pants project contributors (see CONTRIBUTORS.md).
 # Licensed under the Apache License, Version 2.0 (see LICENSE).
-from pants.backend.java.package import deploy_jar  # TODO: Should move to the JVM package.
-from pants.backend.java.target_types import DeployJarTarget  # TODO: Should move to the JVM package.
 from pants.backend.scala.compile import scalac
 from pants.backend.scala.dependency_inference import rules as dep_inf_rules
 from pants.backend.scala.goals import check, repl, tailor
@@ -19,8 +17,9 @@ from pants.backend.scala.test import scalatest
 from pants.jvm import classpath, jdk_rules, resources
 from pants.jvm import util_rules as jvm_util_rules
 from pants.jvm.goals import lockfile
+from pants.jvm.package import deploy_jar
 from pants.jvm.resolve import coursier_fetch, coursier_setup, jvm_tool
-from pants.jvm.target_types import JvmArtifactTarget
+from pants.jvm.target_types import DeployJarTarget, JvmArtifactTarget
 from pants.jvm.test import junit
